requirements.txt
wikipedia-api==0.6.0
sentence-transformers==2.5.1
python-dotenv==1.0.0
flask==3.0.2
requests==2.32.3
beautifulsoup4==4.13.3
pyspellchecker==0.8.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1
googlesearch-python==1.3.0
gunicorn==21.2.0
